28MAY20
Updated class types to distinguish between countries. This in turn allowed me to limit interactions to just Russian-Iranian and Iranian-Iranian. There are no longer any Russian-Russian interactions

Also changed the random choice, allowing agents to interact with any agent they would like, but increasing the amount learned if they interacted with a neighboring agent.

Finally, changed the amount of time each step represented. The previous number was based on how an agent could learn in a consistent day to day interaction over the coruse of the month. The numbers have been reduced so that each day is a step instead.