31MAY20
Seperated the agent classes further, enabling me to expand or limit the number of interactions for each agent class per step. Currently, the model is set up for one step to represent a single month with a capture the flag event hosted by a single Russian agent. The Russian agent is set up to interact with 20 agents per step. Iranian agents are set up to interact with 30 other agents in a month, or roughly 1 per day. Currently, the known issue is that an individual agent does all 30 of its interactions all in a row and then does its whole month worth of self-learning (30 days worth) all at once. This will probably work well enough, but it would good to be able to smooth that out further.